Preemptive multi-tasks.
Call `yield()` in periodly hardware timer interrupt handler to
achieve preemptive multi-tasks. Don't reply on the user task
any more to voluntarily relinquish the CPU.

Additionally, use `sret` to enable interrupts when a user task
starts running. This simulatis the task startup process as if
it returns from a context-switching to unify the code flow.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>