Syscall: sys_exit
Just a simple version now. When the process exits by itself,
the process will be move to state "Zombie" but not be reclaimed
due to we have not supported parent-child mechanism for
processes, will be improved later.

To verify this, type "exit" in the shell, "Exiting shell..." will
be printed and continue press Ctrl+P, will see the state of the
process is marked as 'zombie'.

```shell
1 zombie initcode
```
Signed-off-by: Chen Wang <unicorn_wang@outlook.com>