Syscall: sys_fork
Establish parent-child mechanism for processes, and based on
this to implement the `fork` system call.

Separate the shell logic from `init` and make it a separate
independent program called `sh`. `init` uses `fork` to create
a child process, and then, within that child process, invoke
`exec` to load and execute the `sh` program.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>